Add Notepad, System Info, TCP networking, and desktop management features
- Implemented a Notepad application with basic text editing capabilities, including cursor management and word wrapping.
- Created a System Info window displaying system architecture, video mode, uptime, and memory information.
- Developed a TCP networking stack with connection management, data transmission, and reception functionalities.
- Added desktop management features to handle multiple windows, including creation, focusing, and drawing.
- Introduced a simple browser structure for future web browsing capabilities.
- Implemented a Calculator application structure for basic arithmetic operations.
- Enhanced the desktop interface with a taskbar and start menu for easy access to applications.

+++ kernel/arch/pci.c
@@ -30,8 +30,8 @@ void pci_write_word(uint8 bus, uint8 slot, uint8 func, uint8 offset, uint16 valu
 }
 
 int pci_find_device(uint16 vendor_id, uint16 device_id, pci_device_t* dev) {
-    for (uint8 bus = 0; bus < 256; bus++) {
-        for (uint8 slot = 0; slot < 32; slot++) {
+    for (int bus = 0; bus < 256; bus++) {
+        for (int slot = 0; slot < 32; slot++) {
             uint16 vendor = pci_read_word(bus, slot, 0, PCI_VENDOR_ID);
             if (vendor == 0xFFFF) continue;
